Refactor instrumentors
