+++ src/atla_insights/instrumentation/__init__.py
@@ -1,0 +1,1 @@
+"""Model / LLM framework instrumentation logic."""
