+++ src/atla_insights/instrumentation/_smolagents.py
